requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Azure Functions
azure-functions

# Azure Storage SDK
azure-storage-blob
azure-storage-file-share

# Para manejar los archivos Excel y CSV
pandas
openpyxl

# Para conectarse a SQL Server
pymssql

# (Opcional) Manejo de variables de entorno
python-dotenv